+++ src/main/java/com/atlauncher/data/json/Java.java
@@ -21,8 +21,8 @@ import com.atlauncher.annot.Json;
 
 @Json
 public class Java {
-    private final int min = 0;
-    private final int max = 0;
+    private int min = 0;
+    private int max = 0;
 
     public int getMin() {
         return this.min;
@@ -32,8 +32,7 @@ public class Java {
         return this.max;
     }
 
-    public boolean conforms() {
-        int javaVersion = com.atlauncher.utils.Java.getMinecraftJavaVersionNumber();
+    public boolean conforms(int javaVersion) {
         boolean conforms = true;
 
         if (this.min != 0 && javaVersion < this.min) {