+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Issue with PermSize option being used on custom Java paths in some cases [#520]
+- Wrong Java version warning not working and displaying correctly
 
 ### Misc